Move AW4 files
